refactor: filter 분리 및 에러메시지 정확히 수정
