+++ backend/.gitignore
@@ -38,6 +38,7 @@ out/
 
 ### Custom ###
 src/main/resources/application-secret.yml
+src/test/resources/application-test.yml
 db_dev.mv.db
 db_dev.trace.db
